A fresh start!
The history prior to this commit was an aborted side-project for a purpose
I don't even remember anymore, and so I abandoned it before it became
particularly useful.

Part of the problem with the earlier work is that it was quite inconsistent
as to what level of abstraction it was intended to be at: some of the
operations were just thin wrappers around the raw protocol actions, while
other parts had special additions to try to make it easier to use, but
at the expense of being somewhat opinionated about what callers would be
likely to need.

After this point, this library has some different goals:
- Minimal abstraction: the main point is to abstract away the lowest-level
  wire protocol and its major version negotiation mechanism, so that
  a caller can use the same static types regardless of provider version
  and have the protocol version decision made dynamically inside this
  library.
- Beyond abstracting over major protocol versions, there is as little
  additional abstraction as possible. Different callers can have their own
  opinions about which abstractions are worthwhile and which are just
  overhead. Perhaps another library will one day wrap this one with
  something more opinionated for OpenTofu-shaped callers, but that's out
  of scope here.
- Do as little up-front unmarshalling as possible within the constraints
  of the protocol, so that callers can control when potentially-expensive
  unmarshaling happens and skip it completely if they don't need the
  resulting information.
- Primarily concerned with OpenTofu rather than Terraform. Since today
  both share the same provider protocols designed by HashiCorp this is
  really just a naming decision to avoid implying that this codebase is
  endorsed by HashiCorp by using their trademark, but in the long term
  if OpenTofu introduces its own protocol extensions or new protocol
  versions then this library will prioritize following OpenTofu's
  additions, regardless of whether Terraform supports them.

This commit deletes everything except the plugin startup and shutdown code
and updates to the latest versions of the API protobuf schemas and this
library's Go module dependencies. We'll rebuild the useful client
functionality gradually in subsequent commits, following the principles
described above.

This is also now relicensed under MPL 2.0, just because the protocol buffer
schemas are MPL 2.0 licensed and so this makes the overall licensing story
of this repository easier to understand and gel better with OpenTofu's
own licensing.