providertrace: Optional provider client tracing API
The initial motivation for this is to allow callers to access the same
information that OpenTofu would normally show when environment variables
like TF_LOG and TF_LOG_PROVIDER are set, in case that might help with
debugging how a provider is reacting to a particular request.

Future commits might expose additional hooks here to help callers include
provider calls as part of any existing logging or tracing they already
produce.