Start of modeling provider schema information
This slightly grows the previously-identified scope of this library by
introducing built-in cty decoding of type constraints.

We don't want to expose the raw serialization format of type constraints
because it might change in future protocol versions, but it would also be
a lot of work to recreate the whole cty.Type model in this library, so
we'll be pragmatic and just use cty.Type as an abstraction layer since
that matches how OpenTofu itself represents types and so we can assume it
will be sufficient to model types in this protocol for the foreseeable
future.