Split gRPC-plugin-specific stuff from general provider stuff
We are currently considering offering a new provider plugin model that
would use stdio via pipes rather than gRPC over network sockets, and if
so those plugins would support the same overall protocol concepts but with
a somewhat different execution and handshake process.

To plan ahead a little for what might come, we'll name the current
execution model "GRPCPlugin" in our API here and draw an explicit line
between the concerns specific to that plugin model vs. the more conceptual
protocol model that we'd expect all future protocol designs to build on.