GetProviderSchema implementation for protocol 6
We're starting with this one because this operation has possibly the most
complicated result data structure of any in the protocol and so it's a good
test case for exploring how it looks to follow the "interfaces and
iterators" design for modeling the API while minimizing translation
overhead.

There is not yet an equivalent implementation for protocol 5, so using
this with a protocol-5-only provider will immediately panic. It does seem
to basically work with protocol 6 providers, though.