providerops: IsUnimplementedErr function
This encapsulates the rule for recognizing whether an error returned from
one of the tofuprovider.Provider methods represents that the requested
operation is not implemented at all.

It's important to be able to recognize this case because the protocol
design relies on it to be able to introduce new operations that older
providers would not support, and then the caller can potentially perform
a fallback to some other behavior instead of returning an error.