API defs for the remaining provider-scoped functions
This deals with the rest of the whole-provider-scoped functions that
OpenTofu currently uses, with the actual implementations just stubbed out
for now.

This intentionally ignores some parts of the upstream API that OpenTofu
doesn't use yet. We'll add those in only once we understand whether and how
OpenTofu will use them.