Expose the tfplugin5/tfplugin6 packages for external import
The protobuf Go library expects there to be only one Go package for each
source protobuf module in the entire import graph of any program.

Therefore any program that wants to use this module's protocol abstraction
but _also_ needs to access the lower-level protocol stubs will need to do
the latter through this modules's copy of those stubs.

This is not intended as part of the abstraction offered by this library,
but instead as some pragmatism to allow callers that need to operate at
mixed levels of abstraction either permanently or during a transitional
period when migrating from a local abstraction over the major versions.