Prepare this to become an official OpenTofu repository
This started off as a prototype in a personal repository, but the OpenTofu
project is now intending to use it as the starting point for an
officially-supported client library that OpenTofu itself will hopefully
use eventually.

The main point of this commit is to add the notes warning that this is not
yet ready to use. Aside from anything else, this is currently just my
original prototype as-is, and so its design might change significantly as
it becomes a project maintained by the full set of OpenTofu maintainers.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>